backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
DECIMER
py3Dmol
rdkit
numpy
torch
